refactor(site/src/pages/AgentsPage): extract AgentChatPage hooks
Move conversation editing, workspace watching, and right-panel open state out of AgentChatPage into colocated hooks. Behavior is unchanged.
