+++ site/src/pages/AgentsPage/hooks/useConversationEditingState.test.ts
@@ -1,10 +1,10 @@
 import { act, renderHook } from "@testing-library/react";
 import { createRef } from "react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { useConversationEditingState } from "./AgentChatPage";
-import type { ChatMessageInputRef } from "./components/AgentChatInput";
-import type { PendingAttachment } from "./components/ChatPageContent";
-import { draftInputStorageKeyPrefix } from "./utils/draftStorage";
+import type { ChatMessageInputRef } from "../components/AgentChatInput";
+import type { PendingAttachment } from "../components/ChatPageContent";
+import { draftInputStorageKeyPrefix } from "../utils/draftStorage";
+import { useConversationEditingState } from "./useConversationEditingState";
 
 type MockChatInputHandle = {
 	handle: ChatMessageInputRef;
