+++ site/src/pages/AgentsPage/hooks/useConversationEditingState.ts
@@ -1,0 +1,241 @@
+import { useLayoutEffect, useRef, useState } from "react";
+import type { ChatMessagePart } from "#/api/typesGenerated";
+import { isMobileViewport } from "#/utils/mobile";
+import type { ChatMessageInputRef } from "../components/AgentChatInput";
+import type { PendingAttachment } from "../components/ChatPageContent";
+import {
+	draftInputStorageKeyPrefix,
+	type ParsedDraft,
+	parseStoredDraft,
+} from "../utils/draftStorage";
+
+export class BuiltInCommandPendingError extends Error {}
+
+/** @internal Exported for testing. */
+export function useConversationEditingState(deps: {
+	chatID: string | undefined;
+	onSend: (
+		message: string,
+		attachments?: readonly PendingAttachment[],
+		editedMessageID?: number,
+	) => Promise<void>;
+	chatInputRef: React.RefObject<ChatMessageInputRef | null>;
+	inputValueRef: React.RefObject<string>;
+}) {
+	const { chatID, onSend, chatInputRef, inputValueRef } = deps;
+	const draftStorageKey = chatID
+		? `${draftInputStorageKeyPrefix}${chatID}`
+		: null;
+	const [{ editorInitialValue, initialEditorState }, setDraftState] = useState(
+		() => {
+			if (!draftStorageKey) {
+				return { editorInitialValue: "", initialEditorState: undefined };
+			}
+			const draft = parseStoredDraft(localStorage.getItem(draftStorageKey));
+			return {
+				editorInitialValue: draft.text,
+				initialEditorState: draft.editorState,
+			};
+		},
+	);
+	const serializedEditorStateRef = useRef<string | undefined>(
+		initialEditorState,
+	);
+
+	// Monotonic counter to force LexicalComposer remount.
+	const [remountKey, setRemountKey] = useState(0);
+
+	// Sync the ref with the initial draft value so callers that
+	// read inputValueRef.current see the persisted draft. Uses a
+	// layout effect so the value is available before paint.
+	const initialSyncDone = useRef(false);
+	useLayoutEffect(() => {
+		if (!initialSyncDone.current && editorInitialValue) {
+			initialSyncDone.current = true;
+			(inputValueRef as React.MutableRefObject<string>).current =
+				editorInitialValue;
+		}
+	}, [editorInitialValue, inputValueRef]);
+
+	// -- History editing state --
+	const [editingMessageId, setEditingMessageId] = useState<number | null>(null);
+	const [draftBeforeHistoryEdit, setDraftBeforeHistoryEdit] =
+		useState<ParsedDraft | null>(null);
+	const [editingFileBlocks, setEditingFileBlocks] = useState<
+		readonly ChatMessagePart[]
+	>([]);
+
+	const handleEditUserMessage = (
+		messageId: number,
+		text: string,
+		fileBlocks?: readonly ChatMessagePart[],
+	) => {
+		if (editingMessageId === null) {
+			// Read the current serialized editor state from localStorage
+			// (kept up-to-date by handleContentChange) rather than from
+			// the stale initialEditorState React state.
+			const currentEditorState = draftStorageKey
+				? parseStoredDraft(localStorage.getItem(draftStorageKey)).editorState
+				: undefined;
+			setDraftBeforeHistoryEdit({
+				text: inputValueRef.current,
+				editorState: currentEditorState,
+			});
+		}
+		setEditingMessageId(messageId);
+		setDraftState({
+			editorInitialValue: text,
+			initialEditorState: undefined,
+		});
+		serializedEditorStateRef.current = undefined;
+		setRemountKey((k) => k + 1);
+		inputValueRef.current = text;
+		setEditingFileBlocks(fileBlocks ?? []);
+	};
+
+	const handleCancelHistoryEdit = () => {
+		const savedText = draftBeforeHistoryEdit?.text ?? "";
+		const savedState = draftBeforeHistoryEdit?.editorState;
+		setDraftState({
+			editorInitialValue: savedText,
+			initialEditorState: savedState,
+		});
+		serializedEditorStateRef.current = savedState;
+		setRemountKey((k) => k + 1);
+		inputValueRef.current = savedText;
+		setEditingMessageId(null);
+		setDraftBeforeHistoryEdit(null);
+		setEditingFileBlocks([]);
+	};
+
+	// Clears the composer for an in-flight history edit and
+	// returns a rollback function that restores the editing draft
+	// if the send fails.
+	const clearInputForHistoryEdit = (message: string) => {
+		const snapshot = {
+			editorState: serializedEditorStateRef.current,
+			fileBlocks: editingFileBlocks,
+			messageId: editingMessageId,
+		};
+
+		chatInputRef.current?.clear();
+		inputValueRef.current = "";
+		setEditingMessageId(null);
+
+		return () => {
+			setDraftState({
+				editorInitialValue: message,
+				initialEditorState: snapshot.editorState,
+			});
+			serializedEditorStateRef.current = snapshot.editorState;
+			setRemountKey((k) => k + 1);
+			inputValueRef.current = message;
+			setEditingMessageId(snapshot.messageId);
+			setEditingFileBlocks(snapshot.fileBlocks);
+		};
+	};
+
+	// Clears all input and editing state after a successful send.
+	const finalizeSuccessfulSend = (editedMessageID: number | undefined) => {
+		chatInputRef.current?.clear();
+		if (!isMobileViewport()) {
+			chatInputRef.current?.focus();
+		}
+		inputValueRef.current = "";
+		serializedEditorStateRef.current = undefined;
+		if (draftStorageKey) {
+			localStorage.removeItem(draftStorageKey);
+		}
+		if (editedMessageID !== undefined) {
+			setDraftBeforeHistoryEdit(null);
+			setEditingFileBlocks([]);
+		}
+	};
+
+	// Wraps the parent onSend to clear local input/editing state.
+	const handleSendFromInput = async (
+		message: string,
+		attachments?: readonly PendingAttachment[],
+	) => {
+		const editedMessageID =
+			editingMessageId !== null ? editingMessageId : undefined;
+		const sendPromise = onSend(message, attachments, editedMessageID);
+
+		// For history edits, clear input immediately and prepare
+		// a rollback in case the send fails.
+		const rollback =
+			editedMessageID !== undefined
+				? clearInputForHistoryEdit(message)
+				: undefined;
+
+		try {
+			await sendPromise;
+		} catch (error) {
+			if (error instanceof BuiltInCommandPendingError) {
+				return;
+			}
+			rollback?.();
+			throw error;
+		}
+
+		finalizeSuccessfulSend(editedMessageID);
+	};
+
+	const handleContentChange = (
+		content: string,
+		serializedEditorState: string,
+		hasFileReferences: boolean,
+	) => {
+		inputValueRef.current = content;
+		serializedEditorStateRef.current = serializedEditorState;
+
+		// Don't overwrite the persisted draft while editing a history message.
+		// The original draft is saved in React state and should survive a cancel.
+		if (editingMessageId !== null) {
+			return;
+		}
+
+		if (draftStorageKey) {
+			const shouldPersist = content.trim() || hasFileReferences;
+			if (shouldPersist) {
+				try {
+					localStorage.setItem(draftStorageKey, serializedEditorState);
+				} catch {
+					// QuotaExceededError, silently discard the draft.
+				}
+			} else {
+				localStorage.removeItem(draftStorageKey);
+			}
+		}
+	};
+
+	// Separate from handleContentChange, which avoids setState to prevent
+	// per-keystroke re-renders. The loading editor is a different instance
+	// that unmounts on load, so the seed must advance here.
+	const handleLoadingDraftChange = (
+		content: string,
+		serializedEditorState: string,
+		hasFileReferences: boolean,
+	) => {
+		handleContentChange(content, serializedEditorState, hasFileReferences);
+		setDraftState({
+			editorInitialValue: content,
+			initialEditorState: serializedEditorState,
+		});
+	};
+
+	return {
+		inputValueRef,
+		chatInputRef,
+		editorInitialValue,
+		initialEditorState,
+		remountKey,
+		editingMessageId,
+		editingFileBlocks,
+		handleEditUserMessage,
+		handleCancelHistoryEdit,
+		handleSendFromInput,
+		handleContentChange,
+		handleLoadingDraftChange,
+	};
+}
